+++ zulu/__init__.py
@@ -12,7 +12,7 @@ from .__pkg__ import (
 )
 
 
-from .api import now, parse, range, span_range, delta
+from .api import now, create, parse, range, span_range, delta
 from .datetime import Zulu
 from .parser import ParseError, ISO8601, TIMESTAMP
 from .timedelta import Delta