+++ zulu/api.py
@@ -11,3 +11,10 @@ parse = Zulu.parse
 range = Zulu.range
 span_range = Zulu.span_range
 delta = Delta.parse
+
+
+def create(*args, **kargs):
+    """Factory function to create a new :class:`.Zulu` datetime using the same
+    arguments as the class.
+    """
+    return Zulu(*args, **kargs)